+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 CloudCompare Version History
 ============================
 
+v2.12 (???) - (in development)
+----------------------
+- Improvements
+	- Clipping box tool:
+		- former 'contours' renamed 'envelopes' for the sake of clarity
+		- ability to extract the real contours of the points inside each slice (single slice mode or 'repeat' mode)
+			(CC will rasterize the slice and apply the 'contour plot' extraction algorithm)
+
 v2.11 (Anoia) - (in development)
 ----------------------
 
@@ -84,8 +92,6 @@ v2.11 (Anoia) - (in development)
 		- option to select the extracted contour type (LOWER, UPPER or FULL)
 		- 'up' direction is always Z for slices normal to X or Y (the local X, Y or Z directions of the active cross-section)
 		- 'up' direction is X for slices normal to Z
-		- ability to extract the real contours of the points inside each slice (single slice mode or 'repeat' mode)
-			(CC will rasterize the slice and apply the 'contour plot' extraction algorithm)
 	- All parameters should now be properly remembered from one call to the other (during the same session)
 	- The current box/slice position can now be exported (resp. imported) to (resp. from) the clipboard via the 'Advanced' menu
 	- Command line tool: